# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IUpdateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KataLevel": {
				"enum": [
					"Basic",
					"Medium",
					"High"
				],
				"type": "string"
			},
...
			"IKata": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"level": {
						"$ref": "#/components/schemas/KataLevel"
					},
					"intents": {
						"type": "number",
						"format": "double"
					},
					"stars": {
						"type": "number",
						"format": "double"
					},
					"creator": {
						"type": "string"
					},
					"solution": {
						"type": "string"
					},
					"participants": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"description",
					"level",
					"intents",
					"stars",
					"creator",
					"solution",
					"participants"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "intensivebackendopenbootcamp",
		"version": "1.0.0",
		"description": "Intensive course of Open Bootcamp with Node and Express in TypeScript - Backend",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Victor Canas"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "All user o user found by iD",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retreive the Users in the Collection \"Users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Id of user to retreive (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "message informing if deletion was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to delete the Users in the Collection \"Users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of user to delete (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to update a User in the Collection \"Users\" of DB",
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of user to update",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "IUpdateUser interface to set the new values",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateUser"
							}
						}
					}
				}
			}
		},
		"/api/users/katas": {
			"get": {
				"operationId": "GetKatas",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"UsersController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "level",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/KataLevel"
						}
					}
				]
			}
		},
...
		"/api/katas": {
			"get": {
				"operationId": "GetKatas",
				"responses": {
					"200": {
						"description": "All katas o kata found by ID",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retreive the katas in the Collection \"Katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Id of Kata to retreive (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "level",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/KataLevel"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateKata",
				"responses": {
					"200": {
						"description": "message informing if creating was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to create a Kata in the Collection \"Katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Kata interface to create",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IKata"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteKata",
				"responses": {
					"200": {
						"description": "message informing if deletion was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to delete the Katas in the Collection \"Katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of Kata to delete (optional)",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateKata",
				"responses": {
					"200": {
						"description": "message informing if updating was correct",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to update a Kata in the Collection \"Katas\" of DB",
				"tags": [
					"KatasController"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of Kata to update",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Kata interface to update",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IKata"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}